Flyttet over selector for relevante arbeidsforhold. Rettet feil med race condition i props.
