+++ src/ducks/fagsaker.js
@@ -1,9 +1,6 @@
 import { createSelector, createStructuredSelector } from 'reselect';
 import * as Api from '../services/api';
 import { STATUS, doThenDispatch } from '../services/utils';
-import { FaktaavklaringOppholdPeriodeSelector } from './faktaavklaring';
-
-const moment = require('moment');
 
 // Actions
 const OK = 'fagsaker/OK';
@@ -112,34 +109,6 @@ export const ArbeidsforholdeneSelector = createSelector(
     arbeid.inntekt = inntekt.filter(linje => linje.opplysningspliktigID === arbeid.arbeidsgiverID) || [];
     return arbeid;
   }))
-);
-
-/**
- * Arbeidsforhold refererer til organisasjon med arbeidsforholdID. For at komponenten skal kunne vise
- * navn på arbeidsgiver og evt adresse etc må dette flettes inn i arbeidsforhold. Selectoren gjør en map p
- * alle arbeidsforhold og finner relevant organisasjon etter orgnr og setter hele dette objektet inn
- * i arbeidsforholdet dersom det finnes.
- */
-export const RelevanteArbeidsforholdeneSelector = createSelector(
-  state => (state.fagsaker.data.behandlinger ? state.fagsaker.data.behandlinger[0].saksopplysninger.arbeidsforhold : []),
-  state => OrganisasjonerSelector(state),
-  state => FaktaavklaringOppholdPeriodeSelector(state),
-  (arbeidsforhold = [], organisasjoner = [], opphold = {}) => (arbeidsforhold.map(item => {
-    const arbeid = { ...item };
-    arbeid.arbeidsgiver = organisasjoner.find(org => org.orgnr === arbeid.arbeidsgiverID) || {};
-    return arbeid;
-  }))
-    .filter(arbeidsforholdet => {
-      // Til-og-med-periode for et arbeidsforhold kan være undefined og dermed et fortsatt
-      // løpende arbeidsforhold. For at arbeidsforholdet dermed skal komme med i listen,
-      // sett dagens moment() slik at diff blir riktig.
-      const { fom: ansattFom = moment(), tom: ansattTom = moment() } = arbeidsforholdet.ansettelsesPeriode;
-      const { fom: oppholdFom, tom: oppholdTom } = opphold;
-
-      const arbeidsforholdetPreSoknadsPeriode = moment(oppholdFom, 'YYYY-MM-DD').diff(moment(ansattTom, 'YYYY-MM-DD'));
-      const arbeidsforholdetPostSoknadsPeriode = moment(oppholdTom, 'YYYY-MM-DD').diff(moment(ansattFom, 'YYYY-MM-DD'));
-      return (arbeidsforholdetPreSoknadsPeriode <= 0 && arbeidsforholdetPostSoknadsPeriode >= 0);
-    })
 );
 
 /** Finner alle organisasjonsnummer som er listet i arbeidsforhold.
